run ESLint manually
replace lint-action with custom code (inspired to lint-action's code).
this part of the code should work, as it did for lint-action.
we'll find out if we need to create the checks as well, but that is the
only part that failed from lint-action, while the linting and push
already worked as intended.